[Navigaion] - Refactors navigaion classes into navigation module
