+++ base/src/main/java/edu/artic/base/utils/ActivityExtensions.kt
@@ -1,0 +1,19 @@
+package edu.artic.base.utils
+
+import android.app.Activity
+
+/**
+ * @author Sameer Dhakal (Fuzz)
+ */
+
+
+fun Activity.setWindowFlag(bits: Int, on: Boolean) {
+    val win = window
+    val winParams = win.attributes
+    if (on) {
+        winParams.flags = winParams.flags or bits
+    } else {
+        winParams.flags = winParams.flags and bits.inv()
+    }
+    win.attributes = winParams
+}
