+++ base/src/main/java/edu/artic/base/utils/IntentExtensions.kt
@@ -1,0 +1,19 @@
+package edu.artic.base.utils
+
+import android.content.Intent
+import android.net.Uri
+
+/**
+ *@author Sameer Dhakal (Fuzz)
+ */
+
+/**
+ * Use this to create intents for deep-linking.
+ *
+ * E.g.
+ * `"edu.artic.home".asDeepLinkIntent()`
+ *
+ */
+fun String.asDeepLinkIntent(action: String = Intent.ACTION_VIEW, schema: String = "artic"): Intent {
+    return Intent(action, Uri.parse("$schema://${this}"))
+}
