+++ info/src/main/java/edu/artic/info/InfoActivity.kt
@@ -1,7 +1,7 @@
 package edu.artic.info
 
 import android.os.Bundle
-import edu.artic.base.utils.NavigationSelectListener
+import edu.artic.navigation.NavigationSelectListener
 import edu.artic.base.utils.disableShiftMode
 import edu.artic.ui.BaseActivity
 import kotlinx.android.synthetic.main.activity_info.*
