+++ navigation/build.gradle
@@ -1,0 +1,9 @@
+apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
+
+dependencies {
+    api libs.design_support
+    implementation libs.glide
+    implementation project(':base')
+    implementation libs.constraint_layout
+}
