+++ navigation/src/main/java/edu/artic/navigation/NavigationConstants.kt
@@ -1,0 +1,14 @@
+package edu.artic.navigation
+
+/**
+ *@author Sameer Dhakal (Fuzz)
+ */
+
+class NavigationConstants {
+    companion object {
+        const val HOME: String = "edu.artic.home"
+        const val MAP: String = ""
+        const val AUDIO: String = ""
+        const val INFO: String = "edu.artic.info"
+    }
+}
