+++ settings.gradle
@@ -1,4 +1,4 @@
-include ':app',
+include ':app', ':navigation',
         ':ui',
         ':viewmodel',
         ':base',
