+++ welcome/build.gradle
@@ -21,4 +21,5 @@ dependencies {
     implementation project(':tours')
     implementation project(':exhibitions')
     implementation project(':events')
+    implementation project(':navigation')
 }
